split tests into their own directory
