+++ .npmignore
@@ -3,17 +3,21 @@ test
 tsconfig.json
 tsconfig.module.json
 tslint.json
-.travis.yml
+.idea
 .github
+.nvmrc
 .prettierignore
+.travis.yml
 .vscode
-.circleci
+.whitesource
 build/docs
 **/*.spec.*
 coverage
 .nyc_output
 *.log
 *.lock
-jest.config.json
 .prettierrc.js
 .eslintrc.js
+renovate.json
+CHANGELOG.md
+jest.config.js
