+++ jest.config.js
@@ -125,7 +125,7 @@ module.exports = {
   // roots: [
   //   "<rootDir>"
   // ],
-  roots: ["src"],
+  roots: ["src", "tests"],
 
   // Allows you to use a custom runner instead of Jest's default test runner
   // runner: "jest-runner",
